show data room or house in roomdetail page

+++ hotel_management_system/lib/data/model/home_model.dart
@@ -13,7 +13,7 @@ class HomeModel {
     String? roomType;
     List<String>? imageUrls;
     String? description;
-    int? pricePerNight;
+    double? pricePerNight;
     String? status;
     Owner? owner;
